- pray for it works. At least somewhat.

+++ Lords Mobile/Parse1Packet/HTTPSendData.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int HTTPPostData(int k, int x, int y, char *name, char *guild, char *guildf, int clevel, int kills, int vip, int grank, int might, int HasPrisoners, int plevel);
+int HTTPPostData(int k, int x, int y, char *name, char *guild, char *guildf, int clevel, __int64 kills, int vip, int grank, int might, int HasPrisoners, int plevel);
